Port to ROS2 Humble, add interactive clinic scene and docs
- Port tomography and planner to ROS2 Humble (was ROS1 Noetic)
  - PointField constructor updated to keyword args (prototype.py)
  - Config modules loaded via importlib to avoid name collision
  - GTSAM/smoothing libs preloaded with ctypes.CDLL(RTLD_GLOBAL)

- Fix CUDA 12/13 NVRTC compatibility: float16 -> float in kernel preamble

- Upgrade bundled pybind11 headers from 2.11 to 3.0.2 for NumPy 2.x compat

- Add clinic scene support
  - tomography/config/scene_clinic.py: traversability config for clinic.pcd
  - tomography/scripts/run_standalone.py: ROS-free tomography runner
  - planner/scripts/plan_standalone.py: ROS-free planning runner

- Add interactive ROS2 workflow
  - run_ros2_interactive.py: subscribes to /clicked_point, plans on each
    start+end pair, publishes /pct_path and /pct_marker; z coordinate used
    to select the correct floor/slice automatically
  - launch_ros2.sh: launches node + RViz2 together
  - rsc/rviz/pct_ros2.rviz: RViz2 config with point cloud, tomogram,
    path and marker displays

- Add documentation
  - SETUP.md: step-by-step fresh-clone setup guide
  - PARAMETERS.md: full reference for all tunable parameters
  - README.md: updated for ROS2, new scripts, topics, compatibility notes

+++ tomography/scripts/kernels.py
@@ -5,13 +5,13 @@ import cupy as cp
 def utils_point(resolution, n_row, n_col):
     util_preamble = string.Template(
         '''
-        __device__ int getIndexLine(float16 x, float16 center)
+        __device__ int getIndexLine(float x, float center)
         {
             int i = round((x - center) / ${resolution});
             return i;
         }
 
-        __device__ int getIndexMap_1d(float16 x, float16 y, float16 cx, float16 cy)
+        __device__ int getIndexMap_1d(float x, float y, float cx, float cy)
         {
             // Return 1D index of a point (x, y) in a layer
             int idx_x = getIndexLine(x, cx) + ${n_row} / 2;
